// src/app/models/admin/ruleengine.model.ts
export class Policy {
  policyName: any;
  description: any;
  priority: number;
  isActive: any;
  ticketAbbrevation: any;
  intervalTime: any;
  policyRoles: any;
  tat: any;
  isInternal: any;
  apiName: any;
  apiId: any;
  ruleExpression: any;
  cron: any;
  isDeleted: any;
  deleterUserId: number;
  deletionTime: any;
  lastModificationTime: any;
  lastModifierUserId: number;
  creationTime: any;
  creatorUserId: number;
  id: number;
}